+++ javascript/script.js
@@ -1,4 +1,14 @@
-const appStartTime = performance.now();
+const log = (...msg) => {
+  const dt = new Date();
+  const ts = `${dt.getHours().toString().padStart(2, '0')}:${dt.getMinutes().toString().padStart(2, '0')}:${dt.getSeconds().toString().padStart(2, '0')}.${dt.getMilliseconds().toString().padStart(3, '0')}`;
+  console.log(ts, ...msg); // eslint-disable-line no-console
+};
+
+const debug = (...msg) => {
+  const dt = new Date();
+  const ts = `${dt.getHours().toString().padStart(2, '0')}:${dt.getMinutes().toString().padStart(2, '0')}:${dt.getSeconds().toString().padStart(2, '0')}.${dt.getMilliseconds().toString().padStart(3, '0')}`;
+  console.debug(ts, ...msg); // eslint-disable-line no-console
+};
 
 function gradioApp() {
   const elems = document.getElementsByTagName('gradio-app');